requirements.txt
torch==2.6.0
transformers==4.50.1
pillow==11.1.0